requirements.txt
Flask
pandas
python-binance

